consume assembler test

+++ excercises/pthreads/network_simul_packet_loss/src/simulation/ProducerConsumerTest.hpp
@@ -34,7 +34,7 @@ class ProducerConsumerTest {
   /// Delay of consumer to consume a package, negative for max random
   int consumerDelay = 0;
   /// Percentage of probability of losing a package
-  float packetLoss = 0;
+  int packetLoss = 0;
 
  private:
   /// Producer of the simulated network messages
